# Approximate DBSCAN perf fix