--- a/algorithms/linfa-clustering/README.md
+++ b/algorithms/linfa-clustering/README.md
@@ -21,7 +21,7 @@ You can find a roadmap (and a selection of good first issues)
 Implementation choices, algorithmic details and a tutorial can be found 
 [here](https://docs.rs/linfa-clustering).
 
-Check [here](https://github.com/LukeMathWalker/clustering-benchmarks) for extensive benchmarks against `scikit-learn`'s K-means implementation.
+**WARNING:** Currently the Approximated DBSCAN implementation is slower than the normal DBSCAN implementation. Therefore DBSCAN should always be used over Approximated DBSCAN.
 
 ## License
 Dual-licensed to be compatible with the Rust project.