Turns out I didn't compare benchmarks in my previous PR (#178) and introduced a massive perf regression. The issue was with a faulty range query distance parameter for calculating neighbour cells. The distance was set incorrectly in previous implementations and interacted badly with `linfa-nn`, leading to sky-high runtimes despite correct results. The issue has been fixed and now Approximate DBSCAN runs faster than before the `linfa-nn` integration, although it's still slower than normal DBSCAN.

Pre-`linfa-nn` benchmarks:
```
appx_dbscan/appx_dbscan/10
                        time:   [589.38 us 590.73 us 592.31 us]
                        change: [-3.6668% -2.9207% -2.1162%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  1 (1.00%) low severe
  5 (5.00%) high mild
  9 (9.00%) high severe
appx_dbscan/appx_dbscan/100
                        time:   [6.5295 ms 6.5613 ms 6.6008 ms]
                        change: [-1.9789% -1.2880% -0.5680%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe
Benchmarking appx_dbscan/appx_dbscan/1000: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 9.1s, or reduce sample count to 50.
appx_dbscan/appx_dbscan/1000
                        time:   [90.272 ms 90.557 ms 90.885 ms]
                        change: [+78.299% +79.153% +79.922%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe
```
Benchmarks for this branch:
```
appx_dbscan/appx_dbscan/10
                        time:   [44.176 us 44.398 us 44.629 us]
                        change: [-92.588% -92.524% -92.464%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
Benchmarking appx_dbscan/appx_dbscan/100: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.3s, enable flat sampling, or reduce sample count to 60.
appx_dbscan/appx_dbscan/100
                        time:   [1.0550 ms 1.0857 ms 1.1291 ms]
                        change: [-83.856% -83.575% -83.231%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) high mild
  4 (4.00%) high severe
Benchmarking appx_dbscan/appx_dbscan/1000: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.3s, or reduce sample count to 90.
appx_dbscan/appx_dbscan/1000
                        time:   [52.594 ms 52.781 ms 52.976 ms]
                        change: [-42.001% -41.715% -41.441%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
```
Benchmarks in current master are complete garbage (my bad), so I won't show them.