@YuhanLiin 2021-12-04
Since it's a regression fix I want to merge it sooner than later. As long as I didn't mess something up again it should be fine (and if it's not then people can just use normal DBSCAN). I will wait till end of this weekend to merge.